test(vapor): relocate props-destructure regressions to satisfy source-length check
