+++ crates/vize_atelier_vapor/src/generate/context.rs
@@ -1,10 +1,10 @@
 //! Code generation context that tracks state during Vapor code emission.
 
 use super::{
-    destructure::{parse_destructure_bindings, parse_destructure_names},
+    destructure::{parse_destructure_bindings, parse_destructure_names, resolve_props_binding},
     expression,
 };
-use vize_atelier_core::options::{BindingMetadata, BindingType};
+use vize_atelier_core::options::BindingMetadata;
 use vize_carton::{FxHashMap, FxHashSet, String, ToCompactString, camelize, capitalize, cstr};
 use vize_croquis::builtins::is_global_allowed;
 
@@ -244,26 +244,7 @@ impl<'a> GenerateContext<'a> {
             }
         }
 
-        // Destructured props are compiled away from the setup return object, so
-        // template references must read the reactive props object (the render
-        // signature's `$props`), mirroring the vdom compiler's PROPS binding
-        // resolution. Aliased destructures read the original prop key; the
-        // alias map is consulted for both binding kinds because the merged
-        // script bindings can record an aliased local as plain `Props`.
-        if let Some(bindings) = self.binding_metadata
-            && matches!(
-                bindings.bindings.get(name),
-                Some(BindingType::Props | BindingType::PropsAliased)
-            )
-        {
-            let key = bindings
-                .props_aliases
-                .get(name)
-                .map_or(name, |key| key.as_str());
-            return Some(cstr!("$props.{}", key));
-        }
-
-        None
+        resolve_props_binding(self.binding_metadata, name)
     }
 
     pub(super) fn resolve_simple_reference(&self, expr: &str) -> String {
